requirements.txt
streamlit==1.52.0
pandas==2.3.3
pillow==12.0.0
plotly==6.5.0
psycopg2-binary==2.9.9
